+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
+## [1.10.5] - 2022-12-29
+
+### Changed
+
+-   When growing only copy written content to the new buffer and safe some bytes
+
 ## [1.10.4] - 2022-12-27
 
 ### Fixed
@@ -225,7 +231,9 @@ This release brings extensions to the `ValueStringBuilder` API. For `v1.0` the `
 
 -   Initial release
 
-[Unreleased]: https://github.com/linkdotnet/StringBuilder/compare/1.10.4...HEAD
+[Unreleased]: https://github.com/linkdotnet/StringBuilder/compare/1.10.5...HEAD
+
+[1.10.5]: https://github.com/linkdotnet/StringBuilder/compare/1.10.4...1.10.5
 
 [1.10.4]: https://github.com/linkdotnet/StringBuilder/compare/1.10.3...1.10.4